backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow>=2.10
pillow
numpy
